fix: WIP connection configs
